+++ lib/src/chat_theme.dart
@@ -50,6 +50,7 @@ abstract class ChatTheme {
   /// Creates a new chat theme based on provided colors and text styles.
   const ChatTheme({
     required this.attachmentButtonIcon,
+    required this.attachmentButtonMargin,
     required this.backgroundColor,
     required this.dateDividerMargin,
     required this.dateDividerTextStyle,
@@ -60,6 +61,8 @@ abstract class ChatTheme {
     required this.errorIcon,
     required this.inputBackgroundColor,
     required this.inputBorderRadius,
+    this.inputContainerDecoration,
+    required this.inputMargin,
     required this.inputPadding,
     required this.inputInsetsHorizontal,
     required this.inputInsetsVertical,
@@ -72,6 +75,9 @@ abstract class ChatTheme {
     required this.messageInsetsVertical,
     required this.primaryColor,
     required this.receivedEmojiMessageTextStyle,
+    this.receivedMessageBodyBoldTextStyle,
+    this.receivedMessageBodyCodeTextStyle,
+    this.receivedMessageBodyLinkTextStyle,
     required this.receivedMessageBodyTextStyle,
     required this.receivedMessageCaptionTextStyle,
     required this.receivedMessageDocumentIconColor,
@@ -81,8 +87,12 @@ abstract class ChatTheme {
     required this.seenIcon,
     required this.sendButtonIcon,
     this.sendButtonIconSize,
+    required this.sendButtonMargin,
     required this.sendingIcon,
     required this.sentEmojiMessageTextStyle,
+    this.sentMessageBodyBoldTextStyle,
+    this.sentMessageBodyCodeTextStyle,
+    this.sentMessageBodyLinkTextStyle,
     required this.sentMessageBodyTextStyle,
     required this.sentMessageCaptionTextStyle,
     required this.sentMessageDocumentIconColor,
@@ -98,11 +108,14 @@ abstract class ChatTheme {
   /// Icon for select attachment button
   final Widget? attachmentButtonIcon;
 
+  /// Margin of attachment button
+  final EdgeInsets? attachmentButtonMargin;
+
   /// Used as a background color of a chat widget
   final Color backgroundColor;
 
   /// Margin around date dividers
-  final EdgeInsetsGeometry dateDividerMargin;
+  final EdgeInsets dateDividerMargin;
 
   /// Text style of the date dividers
   final TextStyle dateDividerTextStyle;
@@ -128,8 +141,14 @@ abstract class ChatTheme {
   /// Top border radius of the bottom bar where text field is
   final BorderRadius inputBorderRadius;
 
-  /// Insets of the bottom bar where text field is
-  final EdgeInsetsGeometry inputPadding;
+  /// Decoration of the container wrapping the text field
+  final Decoration? inputContainerDecoration;
+
+  /// Outer insets of the bottom bar where text field is
+  final EdgeInsets inputMargin;
+
+  /// Inner insets of the bottom bar where text field is
+  final EdgeInsets inputPadding;
 
   /// Horizontal input insets
   final double inputInsetsHorizontal;
@@ -165,6 +184,18 @@ abstract class ChatTheme {
   /// Text style used for displaying emojis on text messages
   final TextStyle receivedEmojiMessageTextStyle;
 
+  /// Body text style used for displaying bold text on received text messages.
+  /// Default to a bold version of [receivedMessageBodyTextStyle].
+  final TextStyle? receivedMessageBodyBoldTextStyle;
+
+  /// Body text style used for displaying code text on received text messages.
+  /// Defaults to a mono version of [receivedMessageBodyTextStyle].
+  final TextStyle? receivedMessageBodyCodeTextStyle;
+
+  /// Text style used for displaying link text on received text messages.
+  /// Defaults to [receivedMessageBodyTextStyle]
+  final TextStyle? receivedMessageBodyLinkTextStyle;
+
   /// Body text style used for displaying text on different types
   /// of received messages
   final TextStyle receivedMessageBodyTextStyle;
@@ -195,11 +226,26 @@ abstract class ChatTheme {
   /// Size of send button
   final double? sendButtonIconSize;
 
+  /// Margin of send button
+  final EdgeInsets? sendButtonMargin;
+
   /// Icon for message's `sending` status. For the best look use size of 10.
   final Widget? sendingIcon;
 
   /// Text style used for displaying emojis on text messages
   final TextStyle sentEmojiMessageTextStyle;
+
+  /// Body text style used for displaying bold text on sent text messages.
+  /// Defaults to a bold version of [sentMessageBodyTextStyle].
+  final TextStyle? sentMessageBodyBoldTextStyle;
+
+  /// Body text style used for displaying code text on sent text messages.
+  /// Defaults to a mono version of [sentMessageBodyTextStyle].
+  final TextStyle? sentMessageBodyCodeTextStyle;
+
+  /// Text style used for displaying link text on sent text messages.
+  /// Defaults to [sentMessageBodyTextStyle]
+  final TextStyle? sentMessageBodyLinkTextStyle;
 
   /// Body text style used for displaying text on different types
   /// of sent messages
@@ -220,7 +266,7 @@ abstract class ChatTheme {
   final TextStyle sentMessageLinkTitleTextStyle;
 
   /// Padding around status icons
-  final EdgeInsetsGeometry statusIconPadding;
+  final EdgeInsets statusIconPadding;
 
   /// Color used as a background for user avatar if an image is provided.
   /// Visible if the image has some transparent parts.
@@ -247,8 +293,9 @@ class DefaultChatTheme extends ChatTheme {
   /// which extends [ChatTheme]
   const DefaultChatTheme({
     Widget? attachmentButtonIcon,
+    EdgeInsets? attachmentButtonMargin,
     Color backgroundColor = neutral7,
-    EdgeInsetsGeometry dateDividerMargin = const EdgeInsets.only(
+    EdgeInsets dateDividerMargin = const EdgeInsets.only(
       bottom: 32,
       top: 16,
     ),
@@ -272,9 +319,11 @@ class DefaultChatTheme extends ChatTheme {
     BorderRadius inputBorderRadius = const BorderRadius.vertical(
       top: Radius.circular(20),
     ),
-    EdgeInsetsGeometry inputPadding = EdgeInsets.zero,
     double inputInsetsHorizontal = 24,
     double inputInsetsVertical = 20,
+    Decoration? inputContainerDecoration,
+    EdgeInsets inputMargin = EdgeInsets.zero,
+    EdgeInsets inputPadding = const EdgeInsets.fromLTRB(24, 20, 24, 20),
     Color inputTextColor = neutral7,
     Color? inputTextCursorColor,
     InputDecoration inputTextDecoration = const InputDecoration(
@@ -292,6 +341,9 @@ class DefaultChatTheme extends ChatTheme {
     double messageInsetsVertical = 16,
     Color primaryColor = primary,
     TextStyle receivedEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle? receivedMessageBodyBoldTextStyle,
+    TextStyle? receivedMessageBodyCodeTextStyle,
+    TextStyle? receivedMessageBodyLinkTextStyle,
     TextStyle receivedMessageBodyTextStyle = const TextStyle(
       color: neutral0,
       fontSize: 16,
@@ -321,8 +373,12 @@ class DefaultChatTheme extends ChatTheme {
     Widget? seenIcon,
     Widget? sendButtonIcon,
     double? sendButtonIconSize,
+    EdgeInsets? sendButtonMargin,
     Widget? sendingIcon,
     TextStyle sentEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle? sentMessageBodyBoldTextStyle,
+    TextStyle? sentMessageBodyCodeTextStyle,
+    TextStyle? sentMessageBodyLinkTextStyle,
     TextStyle sentMessageBodyTextStyle = const TextStyle(
       color: neutral7,
       fontSize: 16,
@@ -348,8 +404,7 @@ class DefaultChatTheme extends ChatTheme {
       fontWeight: FontWeight.w800,
       height: 1.375,
     ),
-    EdgeInsetsGeometry statusIconPadding =
-        const EdgeInsets.symmetric(horizontal: 4),
+    EdgeInsets statusIconPadding = const EdgeInsets.symmetric(horizontal: 4),
     Color userAvatarImageBackgroundColor = Colors.transparent,
     List<Color> userAvatarNameColors = colors,
     TextStyle userAvatarTextStyle = const TextStyle(
@@ -365,6 +420,7 @@ class DefaultChatTheme extends ChatTheme {
     ),
   }) : super(
           attachmentButtonIcon: attachmentButtonIcon,
+          attachmentButtonMargin: attachmentButtonMargin,
           backgroundColor: backgroundColor,
           dateDividerMargin: dateDividerMargin,
           dateDividerTextStyle: dateDividerTextStyle,
@@ -375,6 +431,8 @@ class DefaultChatTheme extends ChatTheme {
           errorIcon: errorIcon,
           inputBackgroundColor: inputBackgroundColor,
           inputBorderRadius: inputBorderRadius,
+          inputContainerDecoration: inputContainerDecoration,
+          inputMargin: inputMargin,
           inputPadding: inputPadding,
           inputInsetsHorizontal: inputInsetsHorizontal,
           inputInsetsVertical: inputInsetsVertical,
@@ -387,6 +445,9 @@ class DefaultChatTheme extends ChatTheme {
           messageInsetsVertical: messageInsetsVertical,
           primaryColor: primaryColor,
           receivedEmojiMessageTextStyle: receivedEmojiMessageTextStyle,
+          receivedMessageBodyBoldTextStyle: receivedMessageBodyBoldTextStyle,
+          receivedMessageBodyCodeTextStyle: receivedMessageBodyCodeTextStyle,
+          receivedMessageBodyLinkTextStyle: receivedMessageBodyLinkTextStyle,
           receivedMessageBodyTextStyle: receivedMessageBodyTextStyle,
           receivedMessageCaptionTextStyle: receivedMessageCaptionTextStyle,
           receivedMessageDocumentIconColor: receivedMessageDocumentIconColor,
@@ -397,8 +458,12 @@ class DefaultChatTheme extends ChatTheme {
           seenIcon: seenIcon,
           sendButtonIcon: sendButtonIcon,
           sendButtonIconSize: sendButtonIconSize,
+          sendButtonMargin: sendButtonMargin,
           sendingIcon: sendingIcon,
           sentEmojiMessageTextStyle: sentEmojiMessageTextStyle,
+          sentMessageBodyBoldTextStyle: sentMessageBodyBoldTextStyle,
+          sentMessageBodyCodeTextStyle: sentMessageBodyCodeTextStyle,
+          sentMessageBodyLinkTextStyle: sentMessageBodyLinkTextStyle,
           sentMessageBodyTextStyle: sentMessageBodyTextStyle,
           sentMessageCaptionTextStyle: sentMessageCaptionTextStyle,
           sentMessageDocumentIconColor: sentMessageDocumentIconColor,
@@ -421,8 +486,9 @@ class DarkChatTheme extends ChatTheme {
   /// which extends [ChatTheme]
   const DarkChatTheme({
     Widget? attachmentButtonIcon,
+    EdgeInsets? attachmentButtonMargin,
     Color backgroundColor = dark,
-    EdgeInsetsGeometry dateDividerMargin = const EdgeInsets.only(
+    EdgeInsets dateDividerMargin = const EdgeInsets.only(
       bottom: 32,
       top: 16,
     ),
@@ -446,9 +512,11 @@ class DarkChatTheme extends ChatTheme {
     BorderRadius inputBorderRadius = const BorderRadius.vertical(
       top: Radius.circular(20),
     ),
-    EdgeInsetsGeometry inputPadding = EdgeInsets.zero,
     double inputInsetsHorizontal = 24,
     double inputInsetsVertical = 20,
+    Decoration? inputContainerDecoration,
+    EdgeInsets inputMargin = EdgeInsets.zero,
+    EdgeInsets inputPadding = const EdgeInsets.fromLTRB(24, 20, 24, 20),
     Color inputTextColor = neutral7,
     Color? inputTextCursorColor,
     InputDecoration inputTextDecoration = const InputDecoration(
@@ -466,6 +534,9 @@ class DarkChatTheme extends ChatTheme {
     double messageInsetsVertical = 16,
     Color primaryColor = primary,
     TextStyle receivedEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle? receivedMessageBodyBoldTextStyle,
+    TextStyle? receivedMessageBodyCodeTextStyle,
+    TextStyle? receivedMessageBodyLinkTextStyle,
     TextStyle receivedMessageBodyTextStyle = const TextStyle(
       color: neutral7,
       fontSize: 16,
@@ -495,8 +566,12 @@ class DarkChatTheme extends ChatTheme {
     Widget? seenIcon,
     Widget? sendButtonIcon,
     double? sendButtonIconSize,
+    EdgeInsets? sendButtonMargin,
     Widget? sendingIcon,
     TextStyle sentEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle? sentMessageBodyBoldTextStyle,
+    TextStyle? sentMessageBodyCodeTextStyle,
+    TextStyle? sentMessageBodyLinkTextStyle,
     TextStyle sentMessageBodyTextStyle = const TextStyle(
       color: neutral7,
       fontSize: 16,
@@ -522,8 +597,7 @@ class DarkChatTheme extends ChatTheme {
       fontWeight: FontWeight.w800,
       height: 1.375,
     ),
-    EdgeInsetsGeometry statusIconPadding =
-        const EdgeInsets.symmetric(horizontal: 4),
+    EdgeInsets statusIconPadding = const EdgeInsets.symmetric(horizontal: 4),
     Color userAvatarImageBackgroundColor = Colors.transparent,
     List<Color> userAvatarNameColors = colors,
     TextStyle userAvatarTextStyle = const TextStyle(
@@ -539,6 +613,7 @@ class DarkChatTheme extends ChatTheme {
     ),
   }) : super(
           attachmentButtonIcon: attachmentButtonIcon,
+          attachmentButtonMargin: attachmentButtonMargin,
           backgroundColor: backgroundColor,
           dateDividerMargin: dateDividerMargin,
           dateDividerTextStyle: dateDividerTextStyle,
@@ -549,8 +624,8 @@ class DarkChatTheme extends ChatTheme {
           errorIcon: errorIcon,
           inputBackgroundColor: inputBackgroundColor,
           inputBorderRadius: inputBorderRadius,
-          messageInsetsHorizontal: messageInsetsHorizontal,
-          messageInsetsVertical: messageInsetsVertical,
+          inputContainerDecoration: inputContainerDecoration,
+          inputMargin: inputMargin,
           inputPadding: inputPadding,
           inputInsetsHorizontal: inputInsetsHorizontal,
           inputInsetsVertical: inputInsetsVertical,
@@ -559,8 +634,13 @@ class DarkChatTheme extends ChatTheme {
           inputTextDecoration: inputTextDecoration,
           inputTextStyle: inputTextStyle,
           messageBorderRadius: messageBorderRadius,
+          messageInsetsHorizontal: messageInsetsHorizontal,
+          messageInsetsVertical: messageInsetsVertical,
           primaryColor: primaryColor,
           receivedEmojiMessageTextStyle: receivedEmojiMessageTextStyle,
+          receivedMessageBodyBoldTextStyle: receivedMessageBodyBoldTextStyle,
+          receivedMessageBodyCodeTextStyle: receivedMessageBodyCodeTextStyle,
+          receivedMessageBodyLinkTextStyle: receivedMessageBodyLinkTextStyle,
           receivedMessageBodyTextStyle: receivedMessageBodyTextStyle,
           receivedMessageCaptionTextStyle: receivedMessageCaptionTextStyle,
           receivedMessageDocumentIconColor: receivedMessageDocumentIconColor,
@@ -571,8 +651,12 @@ class DarkChatTheme extends ChatTheme {
           seenIcon: seenIcon,
           sendButtonIcon: sendButtonIcon,
           sendButtonIconSize: sendButtonIconSize,
+          sendButtonMargin: sendButtonMargin,
           sendingIcon: sendingIcon,
           sentEmojiMessageTextStyle: sentEmojiMessageTextStyle,
+          sentMessageBodyBoldTextStyle: sentMessageBodyBoldTextStyle,
+          sentMessageBodyCodeTextStyle: sentMessageBodyCodeTextStyle,
+          sentMessageBodyLinkTextStyle: sentMessageBodyLinkTextStyle,
           sentMessageBodyTextStyle: sentMessageBodyTextStyle,
           sentMessageCaptionTextStyle: sentMessageCaptionTextStyle,
           sentMessageDocumentIconColor: sentMessageDocumentIconColor,