+++ lib/src/util.dart
@@ -1,8 +1,10 @@
 import 'dart:math';
 import 'dart:ui';
+
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:intl/intl.dart';
+
 import './models/date_header.dart';
 import './models/emoji_enlargement_behavior.dart';
 import './models/message_spacer.dart';
@@ -158,6 +160,7 @@ List<Object> calculateChatMessages(
       chatMessages.insert(
         0,
         DateHeader(
+          dateTime: DateTime.fromMillisecondsSinceEpoch(message.createdAt!),
           text: customDateHeaderText != null
               ? customDateHeaderText(
                   DateTime.fromMillisecondsSinceEpoch(message.createdAt!),
@@ -175,11 +178,8 @@ List<Object> calculateChatMessages(
     chatMessages.insert(0, {
       'message': message,
       'nextMessageInGroup': nextMessageInGroup,
-      'showName': notMyMessage &&
-          showUserNames &&
-          showName &&
-          getUserName(message.author).isNotEmpty,
-      'showStatus': true,
+      'showName': notMyMessage && showUserNames && showName,
+      'showStatus': message.showStatus ?? true,
     });
 
     if (!nextMessageInGroup) {
@@ -196,12 +196,14 @@ List<Object> calculateChatMessages(
       chatMessages.insert(
         0,
         DateHeader(
+          dateTime:
+              DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
           text: customDateHeaderText != null
               ? customDateHeaderText(
-                  DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
+                  DateTime.fromMillisecondsSinceEpoch(nextMessage.createdAt!),
                 )
               : getVerboseDateTimeRepresentation(
-                  DateTime.fromMillisecondsSinceEpoch(nextMessage!.createdAt!),
+                  DateTime.fromMillisecondsSinceEpoch(nextMessage.createdAt!),
                   dateFormat: dateFormat,
                   dateLocale: dateLocale,
                   timeFormat: timeFormat,
@@ -212,7 +214,7 @@ List<Object> calculateChatMessages(
 
     if (message is types.ImageMessage) {
       if (kIsWeb) {
-        if (message.uri.startsWith('http')) {
+        if (message.uri.startsWith('http') || message.uri.startsWith('blob')) {
           gallery.add(PreviewImage(id: message.id, uri: message.uri));
         }
       } else {