+++ lib/src/widgets/chat.dart
@@ -1,9 +1,11 @@
 import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:flutter_chat_ui/src/widgets/inherited_l10n.dart';
 import 'package:intl/intl.dart';
 import 'package:photo_view/photo_view_gallery.dart';
+
 import '../chat_l10n.dart';
 import '../chat_theme.dart';
 import '../conditional/conditional.dart';
@@ -11,6 +13,7 @@ import '../models/date_header.dart';
 import '../models/emoji_enlargement_behavior.dart';
 import '../models/message_spacer.dart';
 import '../models/preview_image.dart';
+import '../models/preview_tap_options.dart';
 import '../models/send_button_visibility_mode.dart';
 import '../util.dart';
 import 'chat_list.dart';
@@ -25,12 +28,14 @@ class Chat extends StatefulWidget {
   /// Creates a chat widget
   const Chat({
     Key? key,
+    this.avatarBuilder,
     this.bubbleBuilder,
     this.customBottomWidget,
     this.customDateHeaderText,
     this.customMessageBuilder,
     this.customHeaderTag,
     this.dateFormat,
+    this.dateHeaderBuilder,
     this.dateHeaderThreshold = 900000,
     this.dateLocale,
     this.disableImageGallery,
@@ -42,19 +47,28 @@ class Chat extends StatefulWidget {
     this.imageMessageBuilder,
     this.isAttachmentUploading,
     this.isLastPage,
+    this.isTextMessageTextSelectable = true,
     this.l10n = const ChatL10nEn(),
     required this.messages,
+    this.nameBuilder,
     this.onAttachmentPressed,
+    this.onAvatarTap,
     this.onBackgroundTap,
     this.onEndReached,
     this.onEndReachedThreshold,
+    this.onMessageDoubleTap,
     this.onMessageLongPress,
+    this.onMessageStatusLongPress,
+    this.onMessageStatusTap,
     this.onMessageTap,
+    this.onMessageVisibilityChanged,
     this.onPreviewDataFetched,
     required this.onSendPressed,
     this.onTextChanged,
     this.onTextFieldTap,
     this.sendButtonIsInsideInputBackround = false,
+    this.previewTapOptions = const PreviewTapOptions(),
+    this.scrollController,
     this.scrollPhysics,
     this.sendButtonVisibilityMode = SendButtonVisibilityMode.editing,
     this.showUserAvatars = false,
@@ -65,6 +79,9 @@ class Chat extends StatefulWidget {
     this.usePreviewData = true,
     required this.user,
   }) : super(key: key);
+
+  /// See [Message.avatarBuilder]
+  final Widget Function(String userId)? avatarBuilder;
 
   /// See [Message.bubbleBuilder]
   final Widget Function(
@@ -99,6 +116,9 @@ class Chat extends StatefulWidget {
   /// make sure you initialize your [DateFormat] with a locale. See [customDateHeaderText]
   /// for more customization.
   final DateFormat? dateFormat;
+
+  /// Custom date header builder gives ability to customize date header widget
+  final Widget Function(DateHeader)? dateHeaderBuilder;
 
   /// Time (in ms) between two messages when we will render a date header.
   /// Default value is 15 minutes, 900000 ms. When time between two messages
@@ -144,6 +164,9 @@ class Chat extends StatefulWidget {
   /// See [ChatList.isLastPage]
   final bool? isLastPage;
 
+  /// See [Message.isTextMessageTextSelectable]
+  final bool isTextMessageTextSelectable;
+
   /// Localized copy. Extend [ChatL10n] class to create your own copy or use
   /// existing one, like the default [ChatL10nEn]. You can customize only
   /// certain properties, see more here [ChatL10nEn].
@@ -152,8 +175,14 @@ class Chat extends StatefulWidget {
   /// List of [types.Message] to render in the chat widget
   final List<types.Message> messages;
 
+  /// See [Message.nameBuilder]
+  final Widget Function(String userId)? nameBuilder;
+
   /// See [Input.onAttachmentPressed]
   final void Function()? onAttachmentPressed;
+
+  /// See [Message.onAvatarTap]
+  final void Function(types.User)? onAvatarTap;
 
   /// Called when user taps on background
   final void Function()? onBackgroundTap;
@@ -164,11 +193,24 @@ class Chat extends StatefulWidget {
   /// See [ChatList.onEndReachedThreshold]
   final double? onEndReachedThreshold;
 
+  /// See [Message.onMessageDoubleTap]
+  final void Function(BuildContext context, types.Message)? onMessageDoubleTap;
+
   /// See [Message.onMessageLongPress]
-  final void Function(types.Message)? onMessageLongPress;
+  final void Function(BuildContext context, types.Message)? onMessageLongPress;
+
+  /// See [Message.onMessageStatusLongPress]
+  final void Function(BuildContext context, types.Message)?
+      onMessageStatusLongPress;
+
+  /// See [Message.onMessageStatusTap]
+  final void Function(BuildContext context, types.Message)? onMessageStatusTap;
 
   /// See [Message.onMessageTap]
-  final void Function(types.Message)? onMessageTap;
+  final void Function(BuildContext context, types.Message)? onMessageTap;
+
+  /// See [Message.onMessageVisibilityChanged]
+  final void Function(types.Message, bool visible)? onMessageVisibilityChanged;
 
   /// See [Message.onPreviewDataFetched]
   final void Function(types.TextMessage, types.PreviewData)?
@@ -182,6 +224,12 @@ class Chat extends StatefulWidget {
 
   /// See [Input.onTextFieldTap]
   final void Function()? onTextFieldTap;
+
+  /// See [Message.previewTapOptions]
+  final PreviewTapOptions previewTapOptions;
+
+  /// See [ChatList.scrollController]
+  final ScrollController? scrollController;
 
   /// See [ChatList.scrollPhysics]
   final ScrollPhysics? scrollPhysics;
@@ -299,8 +347,9 @@ class _ChatState extends State<Chat> {
             pageController: PageController(initialPage: _imageViewIndex),
             scrollPhysics: const ClampingScrollPhysics(),
           ),
-          Positioned(
-            right: 16,
+          Positioned.directional(
+            end: 16,
+            textDirection: Directionality.of(context),
             top: 56,
             child: CloseButton(
               color: Colors.white,
@@ -331,14 +380,18 @@ class _ChatState extends State<Chat> {
 
   Widget _messageBuilder(Object object, BoxConstraints constraints) {
     if (object is DateHeader) {
-      return Container(
-        alignment: Alignment.center,
-        margin: widget.theme.dateDividerMargin,
-        child: Text(
-          object.text,
-          style: widget.theme.dateDividerTextStyle,
-        ),
-      );
+      if (widget.dateHeaderBuilder != null) {
+        return widget.dateHeaderBuilder!(object);
+      } else {
+        return Container(
+          alignment: Alignment.center,
+          margin: widget.theme.dateDividerMargin,
+          child: Text(
+            object.text,
+            style: widget.theme.dateDividerTextStyle,
+          ),
+        );
+      }
     } else if (object is MessageSpacer) {
       return SizedBox(
         height: object.height,
@@ -353,6 +406,7 @@ class _ChatState extends State<Chat> {
 
       return Message(
         key: ValueKey(message.id),
+        avatarBuilder: widget.avatarBuilder,
         bubbleBuilder: widget.bubbleBuilder,
         customMessageBuilder: widget.customMessageBuilder,
         customHeaderTag: widget.customHeaderTag,
@@ -360,18 +414,26 @@ class _ChatState extends State<Chat> {
         fileMessageBuilder: widget.fileMessageBuilder,
         hideBackgroundOnEmojiMessages: widget.hideBackgroundOnEmojiMessages,
         imageMessageBuilder: widget.imageMessageBuilder,
+        isTextMessageTextSelectable: widget.isTextMessageTextSelectable,
         message: message,
         messageWidth: _messageWidth,
+        nameBuilder: widget.nameBuilder,
+        onAvatarTap: widget.onAvatarTap,
+        onMessageDoubleTap: widget.onMessageDoubleTap,
         onMessageLongPress: widget.onMessageLongPress,
-        onMessageTap: (tappedMessage) {
+        onMessageStatusLongPress: widget.onMessageStatusLongPress,
+        onMessageStatusTap: widget.onMessageStatusTap,
+        onMessageTap: (context, tappedMessage) {
           if (tappedMessage is types.ImageMessage &&
               widget.disableImageGallery != true) {
             _onImagePressed(tappedMessage);
           }
 
-          widget.onMessageTap?.call(tappedMessage);
+          widget.onMessageTap?.call(context, tappedMessage);
         },
+        onMessageVisibilityChanged: widget.onMessageVisibilityChanged,
         onPreviewDataFetched: _onPreviewDataFetched,
+        previewTapOptions: widget.previewTapOptions,
         roundBorder: map['nextMessageInGroup'] == true,
         showAvatar: map['nextMessageInGroup'] == false,
         showName: map['showName'] == true,
@@ -450,6 +512,7 @@ class _ChatState extends State<Chat> {
                                   onEndReached: widget.onEndReached,
                                   onEndReachedThreshold:
                                       widget.onEndReachedThreshold,
+                                  scrollController: widget.scrollController,
                                   scrollPhysics: widget.scrollPhysics,
                                 ),
                               ),