+++ lib/src/widgets/chat_list.dart
@@ -1,6 +1,8 @@
 import 'package:diffutil_dart/diffutil.dart';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+
 import 'inherited_chat_theme.dart';
 import 'inherited_user.dart';
 
@@ -14,6 +16,7 @@ class ChatList extends StatefulWidget {
     required this.items,
     this.onEndReached,
     this.onEndReachedThreshold,
+    this.scrollController,
     this.scrollPhysics,
   }) : super(key: key);
 
@@ -39,6 +42,9 @@ class ChatList extends StatefulWidget {
   /// next page when scrolled through about 3/4 of the available content.
   final double? onEndReachedThreshold;
 
+  /// Used to control the chat list scroll view
+  final ScrollController? scrollController;
+
   /// Determines the physics of the scroll view
   final ScrollPhysics? scrollPhysics;
 
@@ -53,7 +59,7 @@ class _ChatListState extends State<ChatList>
   final GlobalKey<SliverAnimatedListState> _listKey =
       GlobalKey<SliverAnimatedListState>();
   late List<Object> _oldData = List.from(widget.items);
-  final _scrollController = ScrollController();
+  late ScrollController _scrollController;
 
   late final AnimationController _controller = AnimationController(vsync: this);
 
@@ -66,6 +72,7 @@ class _ChatListState extends State<ChatList>
   void initState() {
     super.initState();
 
+    _scrollController = widget.scrollController ?? ScrollController();
     didUpdateWidget(widget);
   }
 
@@ -183,6 +190,8 @@ class _ChatListState extends State<ChatList>
 
   @override
   Widget build(BuildContext context) {
+    final _query = MediaQuery.of(context);
+
     return NotificationListener<ScrollNotification>(
       onNotification: (notification) {
         if (widget.onEndReached == null || widget.isLastPage == true) {
@@ -228,8 +237,8 @@ class _ChatListState extends State<ChatList>
             ),
           ),
           SliverPadding(
-            padding: const EdgeInsets.only(
-              top: 16,
+            padding: EdgeInsets.only(
+              top: 16 + (kIsWeb ? 0 : _query.padding.top),
             ),
             sliver: SliverToBoxAdapter(
               child: SizeTransition(