+++ lib/src/widgets/message.dart
@@ -1,12 +1,17 @@
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+import 'package:visibility_detector/visibility_detector.dart';
+
 import '../models/emoji_enlargement_behavior.dart';
+import '../models/preview_tap_options.dart';
 import '../util.dart';
 import 'file_message.dart';
 import 'image_message.dart';
 import 'inherited_chat_theme.dart';
 import 'inherited_user.dart';
 import 'text_message.dart';
+import 'user_avatar.dart';
 
 /// Base widget for all message types in the chat. Renders bubbles around
 /// messages and status. Sets maximum width for a message for
@@ -15,6 +20,7 @@ class Message extends StatelessWidget {
   /// Creates a particular message from any message type
   const Message({
     Key? key,
+    this.avatarBuilder,
     this.bubbleBuilder,
     this.customMessageBuilder,
     this.customHeaderTag,
@@ -22,11 +28,19 @@ class Message extends StatelessWidget {
     this.fileMessageBuilder,
     required this.hideBackgroundOnEmojiMessages,
     this.imageMessageBuilder,
+    required this.isTextMessageTextSelectable,
     required this.message,
     required this.messageWidth,
+    this.nameBuilder,
+    this.onAvatarTap,
+    this.onMessageDoubleTap,
     this.onMessageLongPress,
+    this.onMessageStatusLongPress,
+    this.onMessageStatusTap,
     this.onMessageTap,
+    this.onMessageVisibilityChanged,
     this.onPreviewDataFetched,
+    required this.previewTapOptions,
     required this.roundBorder,
     required this.showAvatar,
     required this.showName,
@@ -35,6 +49,10 @@ class Message extends StatelessWidget {
     this.textMessageBuilder,
     required this.usePreviewData,
   }) : super(key: key);
+
+  /// This is to allow custom user avatar builder
+  /// By using this we can fetch newest user info based on id
+  final Widget Function(String userId)? avatarBuilder;
 
   /// Customize the default bubble using this function. `child` is a content
   /// you should render inside your bubble, `message` is a current message
@@ -69,21 +87,46 @@ class Message extends StatelessWidget {
   final Widget Function(types.ImageMessage, {required int messageWidth})?
       imageMessageBuilder;
 
+  /// See [TextMessage.isTextMessageTextSelectable]
+  final bool isTextMessageTextSelectable;
+
   /// Any message type
   final types.Message message;
 
   /// Maximum message width
   final int messageWidth;
 
+  /// See [TextMessage.nameBuilder]
+  final Widget Function(String userId)? nameBuilder;
+
+  /// See [UserAvatar.onAvatarTap]
+  final void Function(types.User)? onAvatarTap;
+
+  /// Called when user double taps on any message
+  final void Function(BuildContext context, types.Message)? onMessageDoubleTap;
+
   /// Called when user makes a long press on any message
-  final void Function(types.Message)? onMessageLongPress;
+  final void Function(BuildContext context, types.Message)? onMessageLongPress;
+
+  /// Called when user makes a long press on status icon in any message
+  final void Function(BuildContext context, types.Message)?
+      onMessageStatusLongPress;
+
+  /// Called when user taps on status icon in any message
+  final void Function(BuildContext context, types.Message)? onMessageStatusTap;
 
   /// Called when user taps on any message
-  final void Function(types.Message)? onMessageTap;
+  final void Function(BuildContext context, types.Message)? onMessageTap;
+
+  /// Called when the message's visibility changes
+  final void Function(types.Message, bool visible)? onMessageVisibilityChanged;
 
   /// See [TextMessage.onPreviewDataFetched]
   final void Function(types.TextMessage, types.PreviewData)?
       onPreviewDataFetched;
+
+  /// See [TextMessage.previewTapOptions]
+  final PreviewTapOptions previewTapOptions;
 
   /// Rounds border of the message to visually group messages together.
   final bool roundBorder;
@@ -110,38 +153,10 @@ class Message extends StatelessWidget {
   /// See [TextMessage.usePreviewData]
   final bool usePreviewData;
 
-  Widget _avatarBuilder(BuildContext context) {
-    final color = getUserAvatarNameColor(
-      message.author,
-      InheritedChatTheme.of(context).theme.userAvatarNameColors,
-    );
-    final hasImage = message.author.imageUrl != null;
-    final initials = getUserInitials(message.author);
-
-    return showAvatar
-        ? Container(
-            margin: const EdgeInsets.only(right: 8),
-            child: CircleAvatar(
-              backgroundColor: hasImage
-                  ? InheritedChatTheme.of(context)
-                      .theme
-                      .userAvatarImageBackgroundColor
-                  : color,
-              backgroundImage:
-                  hasImage ? NetworkImage(message.author.imageUrl!) : null,
-              radius: 16,
-              child: !hasImage
-                  ? Text(
-                      initials,
-                      style: InheritedChatTheme.of(context)
-                          .theme
-                          .userAvatarTextStyle,
-                    )
-                  : null,
-            ),
-          )
-        : const SizedBox(width: 40);
-  }
+  Widget _avatarBuilder() => showAvatar
+      ? avatarBuilder?.call(message.author.id) ??
+          UserAvatar(author: message.author, onAvatarTap: onAvatarTap)
+      : const SizedBox(width: 40);
 
   Widget _bubbleBuilder(
     BuildContext context,
@@ -200,9 +215,12 @@ class Message extends StatelessWidget {
             : TextMessage(
                 emojiEnlargementBehavior: emojiEnlargementBehavior,
                 hideBackgroundOnEmojiMessages: hideBackgroundOnEmojiMessages,
+                isTextMessageTextSelectable: isTextMessageTextSelectable,
                 message: textMessage,
                 customHeaderTag: customHeaderTag,
+                nameBuilder: nameBuilder,
                 onPreviewDataFetched: onPreviewDataFetched,
+                previewTapOptions: previewTapOptions,
                 showName: showName,
                 usePreviewData: usePreviewData,
               );
@@ -261,6 +279,7 @@ class Message extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final _query = MediaQuery.of(context);
     final _user = InheritedUser.of(context).user;
     final _currentUserIsAuthor = _user.id == message.author.id;
     final _enlargeEmojis =
@@ -270,31 +289,35 @@ class Message extends StatelessWidget {
                 emojiEnlargementBehavior, message as types.TextMessage);
     final _messageBorderRadius =
         InheritedChatTheme.of(context).theme.messageBorderRadius;
-    final _borderRadius = BorderRadius.only(
-      bottomLeft: Radius.circular(
+    final _borderRadius = BorderRadiusDirectional.only(
+      bottomEnd: Radius.circular(
+        _currentUserIsAuthor
+            ? roundBorder
+                ? _messageBorderRadius
+                : 0
+            : _messageBorderRadius,
+      ),
+      bottomStart: Radius.circular(
         _currentUserIsAuthor || roundBorder ? _messageBorderRadius : 0,
       ),
-      bottomRight: Radius.circular(_currentUserIsAuthor
-          ? roundBorder
-              ? _messageBorderRadius
-              : 0
-          : _messageBorderRadius),
-      topLeft: Radius.circular(_messageBorderRadius),
-      topRight: Radius.circular(_messageBorderRadius),
+      topEnd: Radius.circular(_messageBorderRadius),
+      topStart: Radius.circular(_messageBorderRadius),
     );
 
     return Container(
-      alignment:
-          _currentUserIsAuthor ? Alignment.centerRight : Alignment.centerLeft,
-      margin: const EdgeInsets.only(
+      alignment: _currentUserIsAuthor
+          ? AlignmentDirectional.centerEnd
+          : AlignmentDirectional.centerStart,
+      margin: EdgeInsetsDirectional.only(
         bottom: 4,
-        left: 20,
+        end: kIsWeb ? 0 : _query.padding.right,
+        start: 20 + (kIsWeb ? 0 : _query.padding.left),
       ),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.end,
         mainAxisSize: MainAxisSize.min,
         children: [
-          if (!_currentUserIsAuthor && showUserAvatars) _avatarBuilder(context),
+          if (!_currentUserIsAuthor && showUserAvatars) _avatarBuilder(),
           ConstrainedBox(
             constraints: BoxConstraints(
               maxWidth: messageWidth.toDouble(),
@@ -303,14 +326,28 @@ class Message extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.end,
               children: [
                 GestureDetector(
-                  onLongPress: () => onMessageLongPress?.call(message),
-                  onTap: () => onMessageTap?.call(message),
-                  child: _bubbleBuilder(
-                    context,
-                    _borderRadius,
-                    _currentUserIsAuthor,
-                    _enlargeEmojis,
-                  ),
+                  onDoubleTap: () => onMessageDoubleTap?.call(context, message),
+                  onLongPress: () => onMessageLongPress?.call(context, message),
+                  onTap: () => onMessageTap?.call(context, message),
+                  child: onMessageVisibilityChanged != null
+                      ? VisibilityDetector(
+                          key: Key(message.id),
+                          onVisibilityChanged: (visibilityInfo) =>
+                              onMessageVisibilityChanged!(message,
+                                  visibilityInfo.visibleFraction > 0.1),
+                          child: _bubbleBuilder(
+                            context,
+                            _borderRadius.resolve(Directionality.of(context)),
+                            _currentUserIsAuthor,
+                            _enlargeEmojis,
+                          ),
+                        )
+                      : _bubbleBuilder(
+                          context,
+                          _borderRadius.resolve(Directionality.of(context)),
+                          _currentUserIsAuthor,
+                          _enlargeEmojis,
+                        ),
                 ),
               ],
             ),
@@ -318,7 +355,14 @@ class Message extends StatelessWidget {
           if (_currentUserIsAuthor)
             Padding(
               padding: InheritedChatTheme.of(context).theme.statusIconPadding,
-              child: showStatus ? _statusBuilder(context) : null,
+              child: showStatus
+                  ? GestureDetector(
+                      onLongPress: () =>
+                          onMessageStatusLongPress?.call(context, message),
+                      onTap: () => onMessageStatusTap?.call(context, message),
+                      child: _statusBuilder(context),
+                    )
+                  : null,
             ),
         ],
       ),